% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eksportujKolor.R
\name{eksportujKolor}
\alias{eksportujKolor}
\title{Zapisuje dane 3 zdjęcia do pliku jako obraz kolorowy, który można otworzyć w
zwykłej przeglądarce obrazów}
\usage{
eksportujKolor(zdjecieR, zdjecieG, zdjecieB, plik, min = 0, max = NA_real_)
}
\arguments{
\item{zdjecieR}{zdjęcie do wykorzystania jako czerwona składowa koloru
(obiekt klasy "RasterLayer")}

\item{zdjecieG}{zdjęcie do wykorzystania jako zielona składowa koloru (obiekt
klasy "RasterLayer")}

\item{zdjecieB}{zdjęcie do wykorzystania jako niebieska składowa koloru
(obiekt klasy "RasterLayer")}

\item{plik}{ścieżka zapisu zdjęcia}

\item{min}{patrz sekcja "details"}

\item{max}{patrz sekcja "details"}
}
\description{
Większość przeglądarek obrazów obsługuje jedynie obrazy, które \itemize{
\item posiadają 1 (monochromatyczne) lub 3 (kolorowe) składowe koloru; \item
wszystkie składowe koloru są 8-bitowe (mają wartości całkowite od 0 do 255).
}

Tymczasem obrazy satelitarne wykorzystują typowo bardziej precyzyjne typy
danych (składowa koloru opisywana 16 czy nawet 32 bitami). Powoduje to, że
zdjęcia satelitarne nie dają się odczytać w "zwykłych" przeglądarkach
obrazów.

Ta funkcja umożliwia zapisanie obrazu tak, by każda składowa koloru opisywana
była jednym bajtem, dzięki czemu obraz da się otworzyć w każdej przeglądarce
obrazów.
}
\details{
Przeskalowanie wartości pikseli następuje według wzoru \code{out = round(255
* (in - min) / (max - min))}.

Przy tym wartości poniżej 0 i powyżej 255 zostaną przycięte do, odpowiednio,
0 i 255.

Jeśli parametr \code{max} nie został podany, zostanie mu domyślnie przypisana
maksymalna wartość piksela spośród wszystkich zdjęć wejściowych.
}

