% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eskportujMono.R
\name{eksportujMono}
\alias{eksportujMono}
\title{Zapisuje dane zdjęcie do pliku jako obraz monochromatyczny, który można
otworzyć w zwykłej przeglądarce obrazów}
\usage{
eksportujMono(zdjecie, plik, min = 0, max = NA_real_, paleta = NULL)
}
\arguments{
\item{zdjecie}{zdjęcie do zapisania (obiekt klasy "RasterLayer")}

\item{plik}{ścieżka zapisu zdjęcia}

\item{min}{patrz sekcja "details"}

\item{max}{patrz sekcja "details"}

\item{paleta}{paleta kolorów - patrz pierwszy argument \code{\link[grDevices]{colorRamp}}}
}
\description{
Większość przeglądarek obrazów obsługuje jedynie obrazy, które \itemize{
\item posiadają 1 (monochromatyczne) lub 3 (kolorowe) składowe koloru; \item
wszystkie składowe koloru są 8-bitowe (mają wartości całkowite od 0 do 255).
}

Tymczasem obrazy satelitarne wykorzystują typowo bardziej precyzyjne typy
danych (składowa koloru opisywana 16 czy nawet 32 bitami). Powoduje to, że
zdjęcia satelitarne nie dają się odczytać w "zwykłych" przeglądarkach
obrazów.

Ta funkcja umożliwia zapisanie obrazu tak, by kolor opisywany był jednym
bajtem, dzięki czemu obraz da się otworzyć w każdej przeglądarce obrazów.
}
\details{
Przeskalowanie wartości pikseli następuje według wzoru \code{out = round(255
* (in - min) / (max - min))}.

Przy tym wartości poniżej 0 i powyżej 255 zostaną przycięte do, odpowiednio,
0 i 255.

Jeśli parametr \code{max} nie został podany, zostanie mu domyślnie przypisana
maksymalna wartość piksela na danym zdjęciu.
}

