% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wczytaj Zdjecia.R
\name{wczytajZdjecia}
\alias{wczytajZdjecia}
\title{Wczytuje do R zdjęcia o wskazanej dacie ze wskazanego katalogu}
\usage{
wczytajZdjecia(katalog, data, skorAtm = TRUE)
}
\arguments{
\item{katalog}{katalog, w którym znajdują się zdjęcia do wczytania}

\item{data}{data zdjęć, które mają być wczytan}

\item{skorAtm}{decyduje, czy wczytać zdjęcia skorygowane czy nieskorygowane
atmosferycznie (TRUE/FALSE, domyślnie TRUE)}
}
\value{
lista z wczytanymi zdjęciami indeksowana nazwami pasm
}
\description{
Wczytuje do R zdjęcia o wskazanej dacie ze wskazanego katalogu
}
\details{
Funkcja zakłada, że zdjęcia mają nazwy takie, jakie nadawaje funkcja
\link{pobierzZdjecia}.
}

